+++ CHIP_IO/OverlayManager.py
@@ -1,3 +1,22 @@
+# Copyright (c) 2016 Robert Wolterman
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy of
+# this software and associated documentation files (the "Software"), to deal in
+# the Software without restriction, including without limitation the rights to
+# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+# of the Software, and to permit persons to whom the Software is furnished to do
+# so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 import os
 import shutil
 import time
@@ -79,7 +98,7 @@ def _set_overlay_verify(name, overlay_path, config_path):
         print("Config path already exists! Not moving forward")
         print("config_path: {0}".format(config_path))
         return -1
-    
+
     # MAKE THE CONFIGURATION PATH
     os.makedirs(config_path)
 
@@ -198,5 +217,5 @@ def unload(overlay):
         _LOADED[overlay.upper()] = False
     else:
         raise ValueError("Invalid Overlay name specified! Choose between: I2C1, SPI2, PWM0, CUST")
-        
+